archive/profiles/.dev/requirements.txt
pyyaml
pandas
tabulate
chromadb
scikit-learn
numpy
joblib
loguru
typer
rich
pytest
bash
shutil